refactor(sync): share SWM descendant graph parser
